Fix #115, added `nosave` and `polling` arguments to SDTmonitor.
The former is used when you do not want to generate a hdf5 file and to save images other than the ones needed for the quicklook page (i.e.: `latest_*`).
The latter is useful when you need to monitor a directory that resides on a shared filesystem, in which the files are generated/moved by another machine and therefore no filesystem events are raised in the machine running the monitor.