Fix #120, unset the numpy version on dependencies. (#124)
Since a new numpy version came out fixing the issue, the line fixing the
numpy version to the latest stable in `.travis.yml` was removed.
I'm forcing a merge because tests are still using the bugged `numpy` package and are still failing because of this. As soon as the package gets updated in Travis and Appveyor, the tests would pass again.