Fix #118, improved monitor. (#119)
Merging with tests using the wrong numpy version. As soon as the package gets updated in the automatic builds I will check for possible errors. I am sorry to do this but I need to merge this updates before proceeding with development. I will spawn only one branch to fix the issues afflicting the SDTmonitor.