Fix #129, fixed the wrong doctest on `scan.py`.
The test was passing before a change to `scan.product_path_from_file_name`
was made. I think that the test was conceptually wrong, since without the
fix to the said method, the SDTmonitor was not working properly because
the `product_path_from_file_name` method was using relative paths
causing some path issues.