Fix #135, the SDTmonitor now can provide HTTP access.
Using the flag `--http-server-port`, followed by the desired port, the whole directory where the SDTmonitor is launched is shared across the network and can be accessed via browser by typing the name or the IP of the machine on which it's running, followed by the said port. I.e. launching the monitor with the following command `SDTmonitor --http-server-port 8080` starts a HTTP server that can be accessed via bworser on the local machine typing `localhost:8080`, or `<hostname or IP>:8080` if you are on a different machine.