Fix #135, changed approach. See changelog below.
Now the HTTP server is launched via code and not as a subprocess. This was made in order to handle the incoming requests by granting them access only to `index.html` and to the generated images. A request to any other file is answered with a 404, file not found error. This was made to prevent access to other non SDTmonitor-related files.

Also, the documentation was updated with the latest CLI additions.