Fix #135, the SDTmonitor now can provide HTTP access. (#136)
* Fix #135, the SDTmonitor now can provide HTTP access.

Using the flag `--http-server-port`, followed by the desired port, the whole directory where the SDTmonitor is launched is shared across the network and can be accessed via browser by typing the name or the IP of the machine on which it's running, followed by the said port. I.e. launching the monitor with the following command `SDTmonitor --http-server-port 8080` starts a HTTP server that can be accessed via browser on the local machine typing `localhost:8080`, or `<hostname or IP>:8080` if you are on a different machine.

* Fix #135, fixed some minor issues introduced with the previous commit.

* Fix #135, changed approach. See changelog below.

Now the HTTP server is launched via code and not as a subprocess. This was made in order to handle the incoming requests by granting them access only to `index.html` and to the generated images. A request to any other file is answered with a 404, file not found error. This was made to prevent access to other non SDTmonitor-related files.

Also, the documentation was updated with the latest CLI additions.

* Fix #135, HTTP access restricted to images having the correct extension

Files matching the `latest_X` pattern but having a different extension than the one specified in the configuration file cannot be accessed anymore.