Fix #153, fixed wrong backend name issue.
This is a temporary fix. Since Sardara backend now can support only 1024
or 16000+ bins per channel, this fix is valid. In the future it will
support also other configurations. When the 2048 bins per channel
configuration will be implemented we will have to find another way to
identify SARDARA or XARCOS backend.