Fix #150, suppressed some warnings in monitor logfile.
I think that preprocessing warnings related to a specific fits file
should not be reported in the monitor logfile. I believe this file
should be used to log the actual behavior of the monitor itself. This
commit in fact suppress all the warnings that may come out from the
image processing method.