Fix #180, major SDTmonitor overhaul
A `Monitor` class has been implemented, in which all the logic of
monitoring for new files and providing analysis results via network
has been moved. The `main_monitor` function now has basically the only
purpose of parsing the CLI arguments and passing them down to an
instance of this `Monitor` class, alongside waiting indefinitely for the
process to be killed by the user.
Internally, the `Monitor` class uses the newly implemented `WebServer`
class, which is a hybrid HTTP and WebSocket server written using the
Tornado library (https://www.tornadoweb.org/en/stable/). The WebServer
listens to a single port for both incoming HTTP and WebSocket requests.
It provided the means to leave the `polling from javascript` strategy
behind for a way more efficient approach in which the server notifies
its clients whenever a new image is available.
Also, `test_monitor.py` has been updated to properly test the newly
written `Monitor` class.