Fix #255, enhanced the SDTmonitor web page
The web page now displays a bar at the bottom. The bar shows all the
channels grouped by feed number (LHCP + RHCP). On top of the bar, in the top 90%
of the window, a single pair of channels is displayed fully. The user
can click on a feed/pair and replace the main one on the top of the screen.
I will post some example images in GitHub.