Fix #255, enhanced the SDTmonitor web page (#256)
* Fix #255, enhanced the SDTmonitor web page

The web page now displays a bar at the bottom. The bar shows all the
channels grouped by feed number (LHCP + RHCP). On top of the bar, in the top 90%
of the window, a single pair of channels is displayed fully. The user
can click on a feed/pair and replace the main one on the top of the screen.
I will post some example images in GitHub.

* Fix #255, add changelog

* Fix #255, fixed formatting

* Fix #255, improved subsequent analysis strategy

This improvements allow the web page to be cleaned when going from a
higher number of images to a lower number. Old images are cleaned and
not displayed anymore. The lower bar is dynamically enlarged or shrinked
relatively to the number of images the server sends.

* Fix #255, fixed remote destination for websocket